requirements.txt
pandas
numpy
streamlit
pyyaml
scikit-learn
openpyxl
plotly
